+++ src/source/speed.rs
@@ -9,7 +9,7 @@
 //! - Updates the try_seek function by multiplying the audio position by the factor.
 //!
 //! To speed up a source from sink all you need to do is call the   `set_speed(factor: f32)` function
-//! For example, here is how you speed up your sound by using sink or playing raw.
+//! For example, here is how you speed up your sound by using sink or playing raw:
 //!
 //! ```no_run
 //!# use std::fs::File;
@@ -17,23 +17,26 @@
 //!# use rodio::{Decoder, Sink, OutputStream, source::{Source, SineWave}};
 //!
 //! // Get an output stream handle to the default physical sound device.
-//! // Note that no sound will be played if _stream is dropped.
-//! let (_stream, stream_handle) = OutputStream::try_default().unwrap();
-//! // Load a sound from a file, using a path relative to Cargo.toml
-//! let file = BufReader::new(File::open("assets/music.ogg").unwrap());
+//! // Note that no sound will be played if the _stream is dropped.
+//! let stream_handle = rodio::OutputStreamBuilder::open_default_stream()
+//!         .expect("open default audio stream");
+//! // Load a sound from a file, using a path relative to `Cargo.toml`
+//! let file = BufReader::new(File::open("examples/music.ogg").unwrap());
 //! // Decode that sound file into a source
 //! let source = Decoder::new(file).unwrap();
 //! // Play the sound directly on the device 2x faster
-//! stream_handle.play_raw(source.convert_samples().speed(2.0)).unwrap();
+//! stream_handle.mixer().add(source.convert_samples().speed(2.0));
 //! std::thread::sleep(std::time::Duration::from_secs(5));
-//!
-//! // Here is how you would do it using the sink.
-//!
+//! ```
+//! Here is how you would do it using the sink:
+//!```no_run
+//! use rodio::source::{Source, SineWave};
 //! let source = SineWave::new(440.0)
-//!     .take_duration(std::time::Duration::from_secs_f32(3.25))
-//!     .amplify(0.20);
-//!
-//! let sink = Sink::try_new(&stream_handle).unwrap();
+//!    .take_duration(std::time::Duration::from_secs_f32(20.25))
+//!    .amplify(0.20);
+//! let stream_handle = rodio::OutputStreamBuilder::open_default_stream()
+//!         .expect("open default audio stream");
+//! let sink = rodio::Sink::connect_new(&stream_handle.mixer());
 //! sink.set_speed(2.0);
 //! sink.append(source);
 //! std::thread::sleep(std::time::Duration::from_secs(5));